refactor(legend): rename to consistent naming convention
Rename OrdinalLegend to LegendOrdinal and ScaleLegend to LegendScale
for consistency with component naming standards.
