+++ lib/components/Legend/LegendOrdinal.vue
