+++ lib/components/Legend/LegendScale.vue
